Correction de divers bugs 2

+++ insamedia/app/Http/Controllers/PublicationController.php
@@ -96,21 +96,23 @@ class PublicationController extends Controller
     }
 
     public function aimer(Request $request, $id){
-        $id = intval($id);
+        if($request->session()->has('id')){
+            $id = intval($id);
         
-        $aime = Aimer::where('idpublication', $id)->where('idcompte', $request->session()->get('id'))->first();
-        if($aime == null){
-            $nouveauAimer = new Aimer;
-            $nouveauAimer->idpublication = $id;
-            $nouveauAimer->idcompte = $request->session()->get('id');
-            $nouveauAimer->save();
-
-            if($request->session()->get('id') != Publication::where('id', $id)->value('idcompte')){
-                NotificationController::enregistrerNotification(Utilisateur::where('id', $request->session()->get('id'))->value('pseudo').' a aimé une publication', Publication::where('id', $id)->value('idcompte'), $request->session()->get('id'), 2, $id);
+            $aime = Aimer::where('idpublication', $id)->where('idcompte', $request->session()->get('id'))->first();
+            if($aime == null){
+                $nouveauAimer = new Aimer;
+                $nouveauAimer->idpublication = $id;
+                $nouveauAimer->idcompte = $request->session()->get('id');
+                $nouveauAimer->save();
+    
+                if($request->session()->get('id') != Publication::where('id', $id)->value('idcompte')){
+                    NotificationController::enregistrerNotification(Utilisateur::where('id', $request->session()->get('id'))->value('pseudo').' a aimé une publication', Publication::where('id', $id)->value('idcompte'), $request->session()->get('id'), 2, $id);
+                }
             }
-        }
-        else{
-            Aimer::where('idpublication', $id)->where('idcompte', $request->session()->get('id'))->delete();
+            else{
+                Aimer::where('idpublication', $id)->where('idcompte', $request->session()->get('id'))->delete();
+            }
         }
         return back();
     }
